refactor: extract pending image actions from AgentWorkspace
